% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentweather.R
\name{get_current_temperature}
\alias{get_current_temperature}
\title{Function to get current temperature}
\usage{
get_current_temperature(location, by = "city", save_dir = "")
}
\arguments{
\item{location}{Location input (city, lat/lon, or postal code)}

\item{by}{Method to specify location ("city", "latlon", "postal")}

\item{save_dir}{Optional directory to save result}
}
\value{
A tibble with temperature data
}
\description{
Function to get current temperature
}
