% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_forecast_by_postal_code}
\alias{get_forecast_by_postal_code}
\title{Get forecast by postal code.}
\usage{
get_forecast_by_postal_code(
  postal_code,
  country,
  save_dir = "",
  language = "en",
  unit = "M",
  day = 16
)
}
\arguments{
\item{postal_code}{Post code}

\item{country}{Country name}

\item{save_dir}{The file path to save result}

\item{language}{Language, default as English}

\item{unit}{Available units, default as Metric (Celsius, m/s, mm)}

\item{day}{Forecast days, default as 16}
}
\description{
Get forecast by postal code.
}
