% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentweather.R
\name{process_api_response}
\alias{process_api_response}
\title{Function to process API response}
\usage{
process_api_response(response)
}
\arguments{
\item{response}{API response object}
}
\value{
Parsed JSON response as a data frame
}
\description{
Function to process API response
}
