% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_csv}
\alias{save_csv}
\title{Save a dataframe as CSV file}
\usage{
save_csv(dataframe, dir = "", endpoint = "", params = c())
}
\arguments{
\item{dataframe}{A dataframe to save}

\item{dir}{The file path}

\item{endpoint}{Endpoint name}

\item{params}{A list of useful information include in the saved CSV file name.}
}
\description{
Save a dataframe as CSV file
}
