% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentweather.R
\name{validate_location}
\alias{validate_location}
\title{Function to validate location input}
\usage{
validate_location(location, by)
}
\arguments{
\item{location}{Location input as a string}

\item{by}{Method to specify location ("city", "latlon", "postal")}
}
\description{
Function to validate location input
}
