% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{weather_alert_city}
\alias{weather_alert_city}
\title{Get alerts by city name}
\usage{
weather_alert_city(city, state = NULL, country = NULL, save_dir = "")
}
\arguments{
\item{city}{A city name}

\item{state}{A state name or code, default is NULL}

\item{country}{A country code, default is NULL}

\item{save_dir}{The file path to save result}
}
\description{
Get alerts by city name
}
