% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{weather_alert_id}
\alias{weather_alert_id}
\title{Get alerts by city id}
\usage{
weather_alert_id(city_id, save_dir = "")
}
\arguments{
\item{save_dir}{The file path to save result}

\item{postal_code}{City ID}
}
\description{
Get alerts by city id
}
