% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{weather_alert_lat}
\alias{weather_alert_lat}
\title{Get alerts by lat&lon}
\usage{
weather_alert_lat(lat, lon, save_dir = "")
}
\arguments{
\item{lat}{Latitude}

\item{lon}{Longtitude}

\item{save_dir}{The file path to save result}
}
\description{
Get alerts by lat&lon
}
