% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{weather_alert_postal}
\alias{weather_alert_postal}
\title{Get alerts by postal code}
\usage{
weather_alert_postal(postal_code, country = NULL, save_dir = "")
}
\arguments{
\item{postal_code}{Postal code}

\item{country}{A country code, default is NULL}

\item{save_dir}{The file path to save result}
}
\description{
Get alerts by postal code
}
